feat: add ublue-motd as a standalone script
This should make it possible to share the ublue-motd script as a package for all the images: it should also be extensible through a configuration file so that we dont need to have a separate script for each thing. This is the first release so I just tried to make it equal to the old one.