chore: move ublue-specific packages to packages/ sub-dir (#195)
This will help us organize our COPR builds better.

I've added a README to each sub-dir which is intended to be built in
COPR, and it points to the respective ublue-os COPR project.

For the moment, the `packages` sub-dir builds are actually happening in
staging, we need to update COPR to build them in the new location before
that will be correct.