fix: ensure libvirt group and libvirtdbus user/group are created
The ublue-os-libvirt-workarounds package already handles creating a
libvirt log dir and restore selinux context on that dir.

Given issues seen in main(Bazzite) and uCore, we should add sysusers.d
entries to ensure that the libvirt group and libvirtdbus user/group are
created properly.

This should be safe even if upstream later includes this configuration.
This should also be safe both on Fedora and CentOS based systems, so we
should be able to use this package on ublue-os desktops, as well as
uCore and Cayo.

Note this also renames a few files for consistency in naming. When
naming files which are workarounds for other packages, we should
namespace using our project or org name to ensure we don't clobber
something which may later be added by upstream.

Fixes: ublue-os/main#748
Fixes: ublue-os/ucore#257