feat(aurora): add start-here logo
adds our white logo as the start-here logo that is used in the default configuration of
the KDE Plasma Panel. That value is only set on initial user login of
fedora kinoite, so this fixes the inconsistency that rebasers have the
purple logo instead of the white one.

For consistency with fedora-logos I put it in scalable/{apps,places}