fix(krunner-bazaar): make bazaar optional
KRunner-bazaar works fine with the flatpak.
Currently there would not be a way to not have bazaar rpm as it is a
hard dependency.
This opens the door for potentially having flatpak bazaar
(no rpm installed) + this plugin and doesn't change anything for
bazzite and aurora as they don't rely on this behavior anyway.